requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
backports.csv==1.0.7
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-import-export==1.2.0
django-rest-knox==4.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.3.2
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
idna==2.7
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
monotonic==1.5
Nuitka==0.6.3.1
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.23.4
pefile==2019.4.18
py2exe==0.9.2.2
pycparser==2.19
PyInstaller==3.4
PyJWT==1.7.1
pyodbc==4.0.26
PyQt5==5.9.2
python-dateutil==2.7.5
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2018.7
pywin32-ctypes==0.2.0
PyYAML==5.1.1
requests==2.20.1
sip==4.19.8
six==1.11.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.15.2
xlrd==1.2.0
xlwt==1.3.0
